Refactors and moves this logic into HospitalController
