+++ server/controllers/__init__.py
@@ -1,0 +1,1 @@
+from .hospital_controller import HospitalController
