+++ server/models/hospital.py
@@ -18,25 +18,6 @@ class Hospital(db.Model):
     self.name = name
     self.address = address
 
-  @classmethod
-  def all(cls):
-    return cls.query.all()
-
-  @classmethod
-  def find(cls, id):
-    return cls.query.get_or_404(id)
-
-  @classmethod
-  def create(cls, payload):
-    return cls(payload.get('name'), payload.get('address'))
-
-  @classmethod
-  def update(cls, id, payload):
-    hospital = cls.find(id)
-    hospital.name = payload.get('name')
-    hospital.address = payload.get('address')
-    return hospital
-
   @property
   def to_json(self):
     return {
